Amqp dump fix

See merge request !13

+++ ttproto/utils/packet_dumper.py
@@ -2,34 +2,117 @@ import os
 import six
 import pika
 import shutil
+import signal
 import logging
 from datetime import datetime
-from multiprocessing import Process
-
+import multiprocessing
 from ttproto.utils.amqp_messages import *
-from ttproto.utils.pure_pcapy import Dumper, Pkthdr, DLT_IEEE802_15_4, DLT_RAW
+from ttproto.utils.pure_pcapy import Dumper, Pkthdr, DLT_RAW, DLT_IEEE802_15_4_NOFCS
+
+logger = logging.getLogger(__name__)
+
+
+def launch_amqp_data_to_pcap_dumper(amqp_url=None, amqp_exchange=None, topics=None, dump_dir=None):
+    def signal_int_handler(self, frame):
+        logger.info('got SIGINT, stopping dumper..')
+
+        if pcap_dumper:
+            pcap_dumper.stop()
+
+    signal.signal(signal.SIGINT, signal_int_handler)
+
+    if amqp_url and amqp_exchange:
+        amqp_exchange = amqp_exchange
+        amqp_url = amqp_url
+
+    else:
+        try:
+            amqp_exchange = str(os.environ['AMQP_EXCHANGE'])
+            print('Imported AMQP_EXCHANGE env var: %s' % amqp_exchange)
+        except KeyError as e:
+            amqp_exchange = "amq.topic"
+            print('Cannot retrieve environment variables for AMQP EXCHANGE. Loading default: %s' % amqp_exchange)
+        try:
+            amqp_url = str(os.environ['AMQP_URL'])
+            print('Imported AMQP_URL env var: %s' % amqp_url)
+            p = six.moves.urllib_parse.urlparse(amqp_url)
+            user = p.username
+            server = p.hostname
+            logger.info(
+                "Env variables imported for AMQP connection, User: {0} @ Server: {1} ".format(user,
+                                                                                              server))
+        except KeyError:
+            print('Cannot retrieve environment variables for AMQP connection. Loading defaults..')
+            # load default values
+            amqp_url = "amqp://{0}:{1}@{2}/{3}".format("guest", "guest", "localhost", "/")
+
+    # connection = pika.BlockingConnection(pika.URLParameters(amqp_url))
+    #
+    # # in case its not declared
+    # connection.channel().exchange_declare(exchange=amqp_exchange,
+    #                                       type='topic',
+    #                                       durable=True,
+    #                                       )
+
+    if topics:
+        pcap_amqp_topic_subscriptions = topics
+    else:
+        pcap_amqp_topic_subscriptions = ['data.serial.fromAgent.coap_client_agent',
+                                         'data.serial.fromAgent.coap_server_agent',
+                                         'data.tun.fromAgent.coap_server_agent',
+                                         'data.tun.fromAgent.coap_client_agent',
+                                         ]
+
+    # init pcap_dumper
+    pcap_dumper = AmqpDataPacketDumper(
+        amqp_url=amqp_url,
+        amqp_exchange=amqp_exchange,
+        topics=pcap_amqp_topic_subscriptions,
+        dump_dir=dump_dir,
+    )
+    # start pcap_dumper
+    pcap_dumper.run()
 
 
 class AmqpDataPacketDumper:
     """
     Sniffs data.serial and dumps into pcap file (assumes that frames are DLT_IEEE802_15_4)
-    Sniffs data.tun and dumps into pcap file (assumes that frames are DLT_IEEE802_15_4)
+    Sniffs data.tun and dumps into pcap file (assumes that frames are DLT_RAW)
 
     about pcap header:
-        ts_sec: the date and time when this packet was captured. This value is in seconds since January 1, 1970 00:00:00 GMT; this is also known as a UN*X time_t. You can use the ANSI C time() function from time.h to get this value, but you might use a more optimized way to get this timestamp value. If this timestamp isn't based on GMT (UTC), use thiszone from the global header for adjustments.
-        ts_usec: in regular pcap files, the microseconds when this packet was captured, as an offset to ts_sec. In nanosecond-resolution files, this is, instead, the nanoseconds when the packet was captured, as an offset to ts_sec /!\ Beware: this value shouldn't reach 1 second (in regular pcap files 1 000 000; in nanosecond-resolution files, 1 000 000 000); in this case ts_sec must be increased instead!
-        incl_len: the number of bytes of packet data actually captured and saved in the file. This value should never become larger than orig_len or the snaplen value of the global header.
-        orig_len: the length of the packet as it appeared on the network when it was captured. If incl_len and orig_len differ, the actually saved packet size was limited by snaplen.
+        ts_sec: the date and time when this packet was captured. This value is in seconds since January 1,
+            1970 00:00:00 GMT; this is also known as a UN*X time_t. You can use the ANSI C time() function
+            from time.h to get this value, but you might use a more optimized way to get this timestamp value.
+            If this timestamp isn't based on GMT (UTC), use thiszone from the global header for adjustments.
+
+        ts_usec: in regular pcap files, the microseconds when this packet was captured, as an offset to ts_sec.
+            In nanosecond-resolution files, this is, instead, the nanoseconds when the packet was captured, as
+            an offset to ts_sec
+            /!\ Beware: this value shouldn't reach 1 second (in regular pcap files 1 000 000;
+            in nanosecond-resolution files, 1 000 000 000); in this case ts_sec must be increased instead!
+
+        incl_len: the number of bytes of packet data actually captured and saved in the file. This value should
+            never become larger than orig_len or the snaplen value of the global header.
+
+        orig_len: the length of the packet as it appeared on the network when it was captured. If incl_len and
+            orig_len differ, the actually saved packet size was limited by snaplen.
     """
     COMPONENT_ID = 'capture_dumper_%s' % uuid.uuid1()  # uuid in case several dumpers listening to bus
     DEFAULT_DUMP_DIR = 'tmp'
-    DEFAULT_802154_DUMP_FILENAME = "DLT_IEEE802_15_4.pcap"
+
     DEFAULT_RAWIP_DUMP_FILENAME = "DLT_RAW.pcap"
+    DEFAULT_802154_DUMP_FILENAME = "DLT_IEEE802_15_4_NO_FCS.pcap"
     NETWORK_DUMPS = [DEFAULT_802154_DUMP_FILENAME, DEFAULT_RAWIP_DUMP_FILENAME]
+
+    DEFAULT_RAWIP_DUMP_FILENAME_WR = "DLT_RAW.pcap~"
+    DEFAULT_802154_DUMP_FILENAME_WR = "DLT_IEEE802_15_4_NO_FCS.pcap~"
+    NETWORK_DUMPS_TEMP = [DEFAULT_RAWIP_DUMP_FILENAME_WR, DEFAULT_802154_DUMP_FILENAME_WR]
+
     QUANTITY_MESSAGES_PER_PCAP = 100
 
-    def __init__(self, amqp_url, amqp_exchange, topics, dump_dir):
+    def __init__(self, amqp_url, amqp_exchange, topics, dump_dir=None):
 
+        self.messages_dumped = 0
         self.url = amqp_url
         self.exchange = amqp_exchange
 
@@ -38,11 +121,16 @@ class AmqpDataPacketDumper:
         else:
             self.dump_dir = self.DEFAULT_DUMP_DIR
 
-        # setup connection
-        self.connection = pika.BlockingConnection(pika.URLParameters(self.url))
+        if not os.path.exists(self.dump_dir):
+            os.makedirs(self.dump_dir)
 
-        # queues & default exchange declaration
-        self.messages_dumped = 0
+        # pcap dumpers
+        self.pcap_15_4_dumper = None
+        self.pcap_raw_ip_dumper = None
+        self.dumpers_init()
+
+        # AMQP stuff
+        self.connection = pika.BlockingConnection(pika.URLParameters(self.url))  # queues & default exchange declaration
         self.channel = self.connection.channel()
 
         self.data_queue_name = 'data@%s' % self.COMPONENT_ID
@@ -51,15 +139,18 @@ class AmqpDataPacketDumper:
                                    arguments={'x-max-length': 1000}
                                    )
 
+        # subscribe to data plane channels
         for t in topics:
             self.channel.queue_bind(exchange=self.exchange,
                                     queue=self.data_queue_name,
                                     routing_key=t)
 
-        if not os.path.exists(self.dump_dir):
-            os.makedirs(self.dump_dir)
+        # subscribe to channel where the terminate session message is published
+        self.channel.queue_bind(exchange=self.exchange,
+                                queue=self.data_queue_name,
+                                routing_key='control.session')
 
-        # Hello world message
+        # publish Hello message in bus
         self.channel.basic_publish(
             body=json.dumps({'_type': '%s.info' % self.COMPONENT_ID,
                              'value': '%s is up!' % self.COMPONENT_ID, }
@@ -74,49 +165,101 @@ class AmqpDataPacketDumper:
         self.channel.basic_qos(prefetch_count=1)
         self.channel.basic_consume(self.on_request, queue=self.data_queue_name)
 
-        # pcap dumpers
-        self.pcap_15_4_dumper = None
-        self.pcap_raw_ip_dumper = None
-        self.dumpers_init()
-
     def dumpers_init(self):
+
+        for net_dump_filename in self.NETWORK_DUMPS_TEMP:
+            full_path = os.path.join(self.dump_dir, net_dump_filename)
+            if os.path.isfile(full_path):
+                os.remove(full_path)
+
         self.pcap_15_4_dumper = Dumper(
-            filename=os.path.join(self.dump_dir, self.DEFAULT_802154_DUMP_FILENAME),
+            filename=os.path.join(self.dump_dir, self.DEFAULT_802154_DUMP_FILENAME_WR),
             snaplen=200,
-            network=DLT_IEEE802_15_4
+            network=DLT_IEEE802_15_4_NOFCS
         )
+
         self.pcap_raw_ip_dumper = Dumper(
-            filename=os.path.join(self.dump_dir, self.DEFAULT_RAWIP_DUMP_FILENAME),
+            filename=os.path.join(self.dump_dir, self.DEFAULT_RAWIP_DUMP_FILENAME_WR),
             snaplen=200,
             network=DLT_RAW
         )
 
+    def dump_packet(self, message):
+
+        try:
+            t = time.time()
+            t_s = int(t)
+            t_u_delta = int((t - t_s) * 1000000)
+            if 'serial' in message.interface_name:
+                raw_packet = bytes(message.data)
+                packet_slip = bytes(message.data_slip)
+
+                # lets build pcap header for packet
+                pcap_packet_header = Pkthdr(
+                    ts_sec=t_s,
+                    ts_usec=t_u_delta,
+                    incl_len=len(raw_packet),
+                    orig_len=len(raw_packet),
+                )
+
+                self.pcap_15_4_dumper.dump(pcap_packet_header, raw_packet)
+
+                self.messages_dumped += 1
+
+                shutil.copyfile(
+                    os.path.join(self.dump_dir, self.DEFAULT_802154_DUMP_FILENAME_WR),
+                    os.path.join(self.dump_dir, self.DEFAULT_802154_DUMP_FILENAME)
+                )
+
+            elif 'tun' in message.interface_name:
+                raw_packet = bytes(message.data)
+
+                # lets build pcap header for packet
+                pcap_packet_header = Pkthdr(
+                    ts_sec=t_s,
+                    ts_usec=t_u_delta,
+                    incl_len=len(raw_packet),
+                    orig_len=len(raw_packet),
+                )
+
+                self.pcap_raw_ip_dumper.dump(pcap_packet_header, raw_packet)
+
+                self.messages_dumped += 1
+
+                shutil.copyfile(
+                    os.path.join(self.dump_dir, self.DEFAULT_RAWIP_DUMP_FILENAME_WR),
+                    os.path.join(self.dump_dir, self.DEFAULT_RAWIP_DUMP_FILENAME)
+                )
+
+            else:
+                logger.info('Raw packet not dumped to pcap: ' + repr(message))
+                return
+
+        except Exception as e:
+            logger.error(e)
+
+        print('Messages dumped : ' + str(self.messages_dumped))
+
     def dumps_rotate(self):
-        logging.info('rotating file dumps')
 
         for net_dump_filename in self.NETWORK_DUMPS:
             full_path = os.path.join(self.dump_dir, net_dump_filename)
             if os.path.isfile(full_path):
+                logger.info('rotating file dump: %s' % full_path)
                 shutil.copyfile(
                     full_path,
-                    os.path.join(self.dump_dir, datetime.now().isoformat() + '_' + net_dump_filename),
+                    os.path.join(self.dump_dir, datetime.now().strftime('%Y%m%d_%H%M%S_') + net_dump_filename),
                 )
-            os.remove(full_path)
-
-        self.dumpers_init()
 
     def stop(self):
-        logging.info("Stopping packet dumper..")
+        logger.info("Stopping packet dumper..")
         self.channel.queue_delete(self.data_queue_name)
         self.channel.stop_consuming()
         self.connection.close()
 
     def on_request(self, ch, method, props, body):
-        now = datetime.now()
-        # obj hook so json.loads respects the order of the fields sent -just for visualization purposeses-
-        req_body_dict = json.loads(body.decode('utf-8'), object_pairs_hook=OrderedDict)
+
         ch.basic_ack(delivery_tag=method.delivery_tag)
-        logging.debug("Message sniffed: %s, body: %s" % (json.dumps(req_body_dict), str(body)))
 
         try:
 
@@ -133,6 +276,7 @@ class AmqpDataPacketDumper:
 
             m = Message.from_json(body)
             m.update_properties(**props_dict)
+            logger.info('got event: %s' % type(m))
 
             if isinstance(m, MsgTestingToolTerminate):
                 ch.stop_consuming()
@@ -140,54 +284,30 @@ class AmqpDataPacketDumper:
 
             if isinstance(m, MsgPacketSniffedRaw):
 
-                self.messages_dumped += 1
-                try:
-                    if self.messages_dumped % self.QUANTITY_MESSAGES_PER_PCAP == 0:  # rotate files each X messages dumped
+                self.dump_packet(m)
+
+                try:  # rotate files each X messages dumped
+                    if self.messages_dumped != 0 and self.messages_dumped % self.QUANTITY_MESSAGES_PER_PCAP == 0:
                         self.dumps_rotate()
+                        self.dumpers_init()
+
                 except Exception as e:
-                    logging.error(e)
-
-                try:
-                    if 'serial' in m.interface_name:
-                        raw_packet = bytes(m.data)
-                        packet_slip = bytes(m.data_slip)
-
-                        # lets build pcap header for packet
-                        pcap_packet_header = Pkthdr(
-                            ts_sec=now.second,
-                            ts_usec=now.microsecond,
-                            incl_len=len(raw_packet),
-                            orig_len=len(raw_packet),
-                        )
-                        self.pcap_15_4_dumper.dump(pcap_packet_header, raw_packet)
-
-                    elif 'tun' in m.interface_name:
-                        raw_packet = bytes(m.data)
-
-                        # lets build pcap header for packet
-                        pcap_packet_header = Pkthdr(
-                            ts_sec=now.second,
-                            ts_usec=now.microsecond,
-                            incl_len=len(raw_packet),
-                            orig_len=len(raw_packet),
-                        )
-                        self.pcap_raw_ip_dumper.dump(pcap_packet_header, raw_packet)
-
-                    else:
-                        logging.info('raw packet not dumped to pcap: ' + repr(m))
-
-                except TypeError as e:
-                    logging.error(str(e))
-                    logging.error(repr(m))
+                    logger.error(e)
 
             else:
-                logging.info('drop amqp message: ' + repr(m))
+                logger.info('drop amqp message: ' + repr(m))
+
 
         except NonCompliantMessageFormatError as e:
             print('* * * * * * API VALIDATION ERROR * * * * * * * ')
             print("AMQP MESSAGE LIBRARY COULD PROCESS JSON MESSAGE")
             print('* * * * * * * * * * * * * * * * * * * * * * * * *  \n')
-            raise NonCompliantMessageFormatError("AMQP MESSAGE LIBRARY COULD PROCESS JSON MESSAGE")
+            # raise NonCompliantMessageFormatError("AMQP MESSAGE LIBRARY COULD PROCESS JSON MESSAGE")
+
+        except Exception as e:
+            logger.error(e)
+            req_body_dict = json.loads(body.decode('utf-8'), object_pairs_hook=OrderedDict)
+            logger.error("Message: %s, body: %s" % (json.dumps(req_body_dict), str(body)))
 
     def run(self):
         print("Starting thread listening on the event bus")
@@ -195,63 +315,11 @@ class AmqpDataPacketDumper:
         print('Bye byes!')
 
 
-def amqp_data_packet_dumper():
-    try:
-        AMQP_EXCHANGE = str(os.environ['AMQP_EXCHANGE'])
-        print('Imported AMQP_EXCHANGE env var: %s' % AMQP_EXCHANGE)
-
-    except KeyError as e:
-        AMQP_EXCHANGE = "amq.topic"
-        print('Cannot retrieve environment variables for AMQP EXCHANGE. Loading default: %s' % AMQP_EXCHANGE)
-
-    try:
-        AMQP_URL = str(os.environ['AMQP_URL'])
-        print('Imported AMQP_URL env var: %s' % AMQP_URL)
-
-        p = six.moves.urllib_parse.urlparse(AMQP_URL)
-
-        AMQP_USER = p.username
-        AMQP_SERVER = p.hostname
-
-        logging.info(
-            "Env variables imported for AMQP connection, User: {0} @ Server: {1} ".format(AMQP_USER,
-                                                                                          AMQP_SERVER))
-
-    except KeyError as e:
-
-        print('Cannot retrieve environment variables for AMQP connection. Loading defaults..')
-        # load default values
-        AMQP_URL = "amqp://{0}:{1}@{2}/{3}".format("guest", "guest", "localhost", "/")
-
-    connection = pika.BlockingConnection(pika.URLParameters(AMQP_URL))
-
-    # in case its not declared
-    connection.channel().exchange_declare(exchange=AMQP_EXCHANGE,
-                                          type='topic',
-                                          durable=True,
-                                          )
-
-    # start pcap dumper
-    pcap_amqp_topic_subscriptions = ['data.serial.fromAgent.coap_client_agent',
-                                     'data.serial.fromAgent.coap_server_agent',
-                                     'data.tun.fromAgent.coap_server_agent',
-                                     'data.tun.fromAgent.coap_client_agent',
-                                     ]
-    pcap_dumper = AmqpDataPacketDumper(
-        amqp_url=AMQP_URL,
-        amqp_exchange=AMQP_EXCHANGE,
-        topics=pcap_amqp_topic_subscriptions,
-        dump_dir=None,
-    )
-    # pcap_dumper.start()
-    pcap_dumper.run()
-
-
 if __name__ == '__main__':
 
     logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.DEBUG)
 
-    p = Process(target=amqp_data_packet_dumper, args=())
+    p = multiprocessing.Process(target=launch_amqp_data_to_pcap_dumper(), args=())
     p.start()
     for i in range(1, 1000):
         time.sleep(1)
